requirements.txt
MarkupSafe~=1.0
Jinja2~=2.11.2
werkzeug~=1.0.1
wsproto~=0.15.0
nonebot[scheduler]==1.8.0
aiocqhttp~=1.4.0
aiohttp~=3.8.1
lxml>=4.9.1
pytz>=2019.3
requests>=2.23.0
sogou_tr_free~=0.0.8
zhconv~=1.4.1
Pillow~=9.1.0 
matplotlib~=3.2.1
numpy~=1.22.3
beautifulsoup4~=4.9.0
pygtrie~=2.3.3
tinydb~=4.1.1
peony-twitter[all]~=2.1.2
cloudscraper~=1.2.60
